snowplow_json_to_postgres_loader/requirements.txt
requests
aws-psycopg2
boto3
pytest